scripts/requirements.txt
newspaper3k
selenium
beautifulsoup4
lxml
ftfy
requests

